+++ source/isaaclab_tasks/isaaclab_tasks/core/reorient/config/allegro_hand/allegro_hand_direct_env_cfg.py
@@ -3,142 +3,27 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-
-from isaaclab_newton.physics import MJWarpSolverCfg, NewtonCfg
-from isaaclab_ovphysx.physics import OvPhysxCfg
-from isaaclab_physx.physics import PhysxCfg
-
-import isaaclab.sim as sim_utils
-from isaaclab.assets import ArticulationCfg, RigidObjectCfg
+from isaaclab.assets import ArticulationCfg
 from isaaclab.envs import DirectRLEnvCfg
 from isaaclab.markers import VisualizationMarkersCfg
 from isaaclab.scene import InteractiveSceneCfg
 from isaaclab.sim import SimulationCfg
 from isaaclab.sim.spawners.materials import RigidBodyMaterialBaseCfg
-from isaaclab.utils.assets import ISAAC_NUCLEUS_DIR
 from isaaclab.utils.configclass import configclass
 
-from isaaclab_tasks.core.reorient.reorient_task_base import (
+from isaaclab_tasks.core.reorient.config.allegro_hand.allegro_hand_common import (
     ALLEGRO_ACTUATED_JOINT_NAMES,
     ALLEGRO_FINGERTIP_BODY_NAMES,
-)
-from isaaclab_tasks.utils import PresetCfg
-
-from isaaclab_assets.robots.allegro import ALLEGRO_HAND_CFG
-
-
-@configclass
-class ObjectCfg(PresetCfg):
-    physx = RigidObjectCfg(
-        prim_path="/World/envs/env_.*/object",
-        spawn=sim_utils.UsdFileCfg(
-            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/Blocks/DexCube/dex_cube_instanceable.usd",
-            rigid_props=sim_utils.RigidBodyPropertiesCfg(
-                kinematic_enabled=False,
-                disable_gravity=False,
-                enable_gyroscopic_forces=True,
-                solver_position_iteration_count=8,
-                solver_velocity_iteration_count=0,
-                sleep_threshold=0.005,
-                stabilization_threshold=0.0025,
-                max_depenetration_velocity=1000.0,
-            ),
-            mass_props=sim_utils.MassPropertiesCfg(density=400.0),
-            scale=(1.2, 1.2, 1.2),
-        ),
-        init_state=RigidObjectCfg.InitialStateCfg(pos=(0.0, -0.17, 0.56), rot=(0.0, 0.0, 0.0, 1.0)),
-    )
-    newton_mjwarp = ArticulationCfg(
-        prim_path="/World/envs/env_.*/object",
-        spawn=sim_utils.UsdFileCfg(
-            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/Blocks/DexCube/dex_cube_instanceable.usd",
-            mass_props=sim_utils.MassPropertiesCfg(density=400.0),
-            scale=(1.2, 1.2, 1.2),
-        ),
-        init_state=ArticulationCfg.InitialStateCfg(
-            pos=(0.0, -0.17, 0.565), rot=(0.0, 0.0, 0.0, 1.0), joint_pos={}, joint_vel={}
-        ),
-        actuators={},
-        articulation_root_prim_path="",
-    )
-    ovphysx = RigidObjectCfg(
-        prim_path="/World/envs/env_.*/object",
-        spawn=sim_utils.UsdFileCfg(
-            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/Blocks/DexCube/dex_cube_instanceable.usd",
-            rigid_props=sim_utils.RigidBodyPropertiesCfg(
-                kinematic_enabled=False,
-                disable_gravity=False,
-                enable_gyroscopic_forces=True,
-                solver_position_iteration_count=8,
-                solver_velocity_iteration_count=0,
-                sleep_threshold=0.005,
-                stabilization_threshold=0.0025,
-                max_depenetration_velocity=1000.0,
-            ),
-            mass_props=sim_utils.MassPropertiesCfg(density=400.0),
-            scale=(1.2, 1.2, 1.2),
-        ),
-        init_state=RigidObjectCfg.InitialStateCfg(pos=(0.0, -0.17, 0.56), rot=(0.0, 0.0, 0.0, 1.0)),
-    )
-    default = physx
-
-
-@configclass
-class PhysicsCfg(PresetCfg):
-    physx = PhysxCfg(
-        bounce_threshold_velocity=0.2,
-    )
-    newton_mjwarp = NewtonCfg(
-        solver_cfg=MJWarpSolverCfg(
-            solver="newton",
-            integrator="implicitfast",
-            njmax=80,
-            nconmax=70,
-            impratio=10.0,
-            cone="elliptic",
-            update_data_interval=2,
-            iterations=100,
-            # save_to_mjcf="AllegroHand.xml",
-        ),
-        num_substeps=2,
-        debug_mode=False,
-    )
-    ovphysx = OvPhysxCfg()
-    default = physx
-
-
-# Scene pieces shared verbatim by the manager-based variant.
-ROBOT_CFG = ALLEGRO_HAND_CFG.replace(prim_path="/World/envs/env_.*/Robot")
-OBJECT_CFG = ObjectCfg()
-GOAL_OBJECT_CFG = VisualizationMarkersCfg(
-    prim_path="/Visuals/goal_marker",
-    markers={
-        "goal": sim_utils.UsdFileCfg(
-            usd_path=f"{ISAAC_NUCLEUS_DIR}/Props/Blocks/DexCube/dex_cube_instanceable.usd",
-            scale=(1.2, 1.2, 1.2),
-        )
-    },
+    GOAL_OBJECT_CFG,
+    OBJECT_CFG,
+    ROBOT_CFG,
+    ObjectCfg,
+    PhysicsCfg,
 )
 
 
-# Simulation settings shared by the Direct and manager variants (configclass
-# deep-copies these defaults per cfg instance). The solver-common base material
-# is sufficient: only friction values are set, so no PhysX-specific
-# ``physxMaterial`` attributes are authored.
 @configclass
-class AllegroHandTaskCfgBase:
-    """Shared Allegro task settings inherited by both the Direct and the manager configurations."""
-
-    sim: SimulationCfg = SimulationCfg(
-        dt=1 / 120,
-        render_interval=4,
-        physics_material=RigidBodyMaterialBaseCfg(static_friction=1.0, dynamic_friction=1.0),
-        physics=PhysicsCfg(),
-    )
-
-
-@configclass
-class AllegroHandEnvCfg(AllegroHandTaskCfgBase, DirectRLEnvCfg):
+class AllegroHandEnvCfg(DirectRLEnvCfg):
     # env
     decimation = 4
     episode_length_s = 10.0
@@ -147,6 +32,14 @@ class AllegroHandEnvCfg(AllegroHandTaskCfgBase, DirectRLEnvCfg):
     state_space = 0
     asymmetric_obs = False
     obs_type = "full"
+
+    # simulation — values mirrored by the manager cfg (guarded by the value-parity test)
+    sim: SimulationCfg = SimulationCfg(
+        dt=1 / 120,
+        render_interval=4,
+        physics_material=RigidBodyMaterialBaseCfg(static_friction=1.0, dynamic_friction=1.0),
+        physics=PhysicsCfg(),
+    )
     # robot
     robot_cfg: ArticulationCfg = ROBOT_CFG
